Updated all the files.
This is mostly working with the current hardware, but there are some problems.

+++ forwarddec.cpp
@@ -14,9 +14,10 @@ void setSequencePriority(const int priority){
 
 namespace interface{
 	namespace pattSwitch{
-		uint16_t selectedPattern = 0;
+		uint32_t selectedPattern = 0;
 	}
 	namespace settings{
 		unsigned int buttonHoldDelay = 250;
 	}
 }
+
